Alterações no arquivo PhotoController.js

+++ backend/controllers/PhotoController.js
@@ -37,9 +37,12 @@ const deletePhoto = async(req,res) =>{
 
     const reqUser = req.user
 
-    const photo = Photo.findById(mongoose.Types.ObjectId(id));
+    // const photo = Photo.findById(mongoose.Types.ObjectId(id));
 
-    //Check if photo exists
+    const photo = Photo.findById(id);
+
+    try {
+        //Check if photo exists
     if (!photo) {
         res.status(404).json({errors:["Foto não encontrada!"]})
      return;
@@ -53,6 +56,33 @@ const deletePhoto = async(req,res) =>{
         });
     }
 
+    await Photo.findByIdAndDelete(photo._id);
+
+        res.
+        status(200).
+        json({id:photo._id,message:"Foto excluida com sucesso."});
+    } catch (error) {
+         res.
+         status(404).
+         json({errors:["Foto não encontrada!"]})
+
+         return;
+    }
+
+    //Check if photo exists
+    if (!photo) {
+        res.status(404).json({errors:["Foto não encontrada!"]})
+     return;
+    }
+
+    //Check if photo belongs to user
+    if (!photo.userId.equals(reqUser._id)) {
+        res
+        .status(422)
+        .json({errors:["Ocorreu um erro,por favor tente novamente mais tarde."]});
+        
+    }
+
     await Photo.findByIdAndDelete(photo._id)
 
     res.status(200).json({id:photo._id,message:"Foto excluida com sucesso."});
@@ -60,4 +90,5 @@ const deletePhoto = async(req,res) =>{
 
 module.exports = {
     insertPhoto,
+    deletePhoto,
 };
